David Surkov PhD thesis Royal Holloway 2004
Abstract
Since the dawn of the computer era everyone has been fascinated about the idea
of creating an Artificial Intelligence. In the early days it seemed like it was
around the corner. Nevertheless, every advance in computer science has been
pushing the creation of AI further and further away. This is mainly attributed to
the fact that we are still not entirely united in defining the meaning and essence
of intelligence itself, let alone the Artificial Intelligence.

Meanwhile, another branch of computer science, called Machine Learning, has
evolved. Researchers in this field have been concentrating on developing computer
algorithms to learn data and to produce output when presented with new
data from the same knowledge domain. This field turned out to be more practical
and therefore useful in the real world. However, little attention has been
drawn into how the advances in Machine Learning can lead to the AI creation.

In this thesis we overview the Machine Learning field to understand its current
role and where it fits within the AI development. We analyse problems currently
experienced in this field and identify what needs to be introduced in
order to advance further in the direction of AI creation.

A typical machine learning algorithm is designed to produce bare predictions.
We show that the next major step in the Machine Learning is the development
of algorithms that produce not only a prediction but also quality measures
whereby the prediction can be assessed rather than unconditionally accepted. In
other terms, we argue that machine learning algorithms can potentially resemble
intelligence if they could produce qualified predictions.

Following the idea of algorithmic intelligence, we introduce and analyse a way
of obtaining qualified predictions based on Algorithmic Information Theory or
Algorithmic Theory of Randomness in particular. By creating so called Confidence
Machine, that is capable of producing measures of confidence and
credibility for each prediction along with the prediction itself, and by demonstrating
its capabilities practically, we make philosophical speculations about
further possible developments of Artificial Intelligence.